CHG: long double time keeping

+++ src/cosima/inc/MCRun.hh
@@ -175,7 +175,7 @@ public:
   MCSource* GetSource(unsigned int i);
 
   /// Add an event to an event list (used for activation simulation)
-  bool AddToBuildUpEventList(double Energy, G4ThreeVector Position, G4ThreeVector Direction, G4ThreeVector Polarization, double Time, G4ParticleDefinition* ParticleType, MString VolumeName);
+  bool AddToBuildUpEventList(double Energy, G4ThreeVector Position, G4ThreeVector Direction, G4ThreeVector Polarization, long double Time, G4ParticleDefinition* ParticleType, MString VolumeName);
 
   /// Skip one event in the activation source given the particle type and volume name 
   bool SkipOneEvent(G4ParticleDefinition* ParticleType, MString VolumeName); 
@@ -188,9 +188,9 @@ public:
   void SaveIsotopeStore();
 
   /// Set the current time
-  void SetSimulatedTime(double Time) { m_SimulatedTime = Time; }
+  void SetSimulatedTime(long double Time) { m_SimulatedTime = Time; }
   /// Set the current time
-  double GetSimulatedTime() const { return m_SimulatedTime; }
+  long double GetSimulatedTime() const { return m_SimulatedTime; }
 
   /// Set the number of generated particles in the particle source
   void SetNGeneratedParticles(int NGeneratedParticles) { m_NGeneratedParticles = NGeneratedParticles; }
@@ -289,7 +289,7 @@ private:
   double m_NextEmissionScale;
 
   /// Duration of this run in seconds
-  double m_Duration;
+  long double m_Duration;
   /// Maximum bnumber of triggered events
   long m_Triggers;
   /// Maximum number of events
@@ -311,7 +311,7 @@ private:
   unsigned int m_NAddedIsotopes;
 
   /// Simulated time
-  double m_SimulatedTime;
+  long double m_SimulatedTime;
 
   /// Number of simulated events
   long m_NSimulatedEvents;
